+++ tenable_io/client.py
@@ -15,6 +15,7 @@ from tenable_io.api.agents import AgentsApi
 from tenable_io.api.assets import AssetsApi
 from tenable_io.api.base import BaseRequest
 from tenable_io.api.bulk_operations import BulkOperationsApi
+from tenable_io.api.credentials import CredentialsApi
 from tenable_io.api.editor import EditorApi
 from tenable_io.api.exclusions import ExclusionApi
 from tenable_io.api.exports import ExportsApi
@@ -110,6 +111,7 @@ class TenableIOClient(object):
         self.agents_api = AgentsApi(self)
         self.assets_api = AssetsApi(self)
         self.bulk_operations_api = BulkOperationsApi(self)
+        self.credentials_api = CredentialsApi(self)
         self.editor_api = EditorApi(self)
         self.exclusions_api = ExclusionApi(self)
         self.exports_api = ExportsApi(self)